Refactor: Remove ConsoleGraphics project and associated graphics classes; integrate graphics functionality into UI namespace
